Updates. Done for a bit.
